Keep shaker API refactor and restore BioShake heating backend
