+++ pylabrobot/shaking/__init__.py
@@ -1,4 +1,3 @@
 from .backend import ShakerBackend
-from .bioshake_backend import BioShake
 from .chatterbox import ShakerChatterboxBackend
 from .shaker import Shaker
